feat: 욕ì욕설 필터링

+++ BE/src/main/java/com/example/b101/service/SceneService.java
@@ -4,6 +4,8 @@ import com.example.b101.cache.Game;
 import com.example.b101.cache.SceneRedis;
 import com.example.b101.common.ApiResponseUtil;
 import com.example.b101.domain.PlayerStatus;
+import com.example.b101.domain.StoryCard;
+import com.example.b101.dto.DeleteSceneRequest;
 import com.example.b101.dto.GenerateSceneRequest;
 import com.example.b101.dto.SceneRequest;
 import com.example.b101.repository.GameRepository;
@@ -32,7 +34,7 @@ public class SceneService {
 
 
 
-    public ResponseEntity<?> createScene(SceneRequest sceneRequest, HttpServletRequest request , boolean isStoryCard) {
+    public ResponseEntity<?> createScene(SceneRequest sceneRequest, HttpServletRequest request) {
         // 게임 데이터 조회 및 유효성 검사
         Game game = gameRepository.findById(sceneRequest.getGameId());
         if (game == null) {
@@ -53,7 +55,7 @@ public class SceneService {
         // GPU 서버 요청을 위한 객체 생성
         GenerateSceneRequest generateSceneRequest = GenerateSceneRequest.builder()
                 .session_id(sceneRequest.getGameId())            // 게임 아이디 (세션 식별자)
-                .game_mode(1)                    // 작화 스타일 (예: 1: 기본 모드)
+                .game_mode(game.getDrawingStyle())                    // 작화 스타일 (예: 1: 기본 모드)
                 .user_sentence(sceneRequest.getUserPrompt()) // 사용자 프롬프트
                 .status(0)                       // 진행 상태 (0: 진행 중)
                 .build();
@@ -70,7 +72,7 @@ public class SceneService {
                     .bodyToMono(byte[].class)
                     .block();
         } catch (WebClientException e) {
-            return ApiResponseUtil.failure("GPU 서버 통신 중 오류 발생 : " + e.getMessage(),
+            return ApiResponseUtil.failure("GPU 서버 통신 중 오류 발생",
                     HttpStatus.INTERNAL_SERVER_ERROR,
                     request.getRequestURI());
         }
@@ -93,13 +95,6 @@ public class SceneService {
 
         redisSceneRepository.save(scene);
 
-        PlayerStatus playerStatus = gameRepository.getPlayerStatus(sceneRequest.getGameId(), sceneRequest.getUserId());
-
-        //사용한 카드 삭제
-        if(isStoryCard) {
-            playerStatus.getStoryCards().removeIf(storyCard -> storyCard.getId() == sceneRequest.getCardId());
-        }
-
 
         // 이미지 바이너리 데이터를 PNG 미디어 타입으로 반환
         return ResponseEntity.status(HttpStatus.CREATED)
@@ -108,18 +103,32 @@ public class SceneService {
     }
 
 
+    public ResponseEntity<?> deleteScene(DeleteSceneRequest deleteSceneRequest, HttpServletRequest request) {
 
-
-    public ResponseEntity<?> deleteScene(String gameId, HttpServletRequest request) {
-        List<SceneRedis> scenes = redisSceneRepository.findAllByGameId(gameId);
+        List<SceneRedis> scenes = redisSceneRepository.findAllByGameId(deleteSceneRequest.getGameId());
         if(scenes.isEmpty()) {
             return ApiResponseUtil.failure("아직 저장된 scene이 없습니다.",
                     HttpStatus.BAD_REQUEST,
                     request.getRequestURI());
         }
-        SceneRedis lastScene = scenes.get(scenes.size() - 1);
-        redisSceneRepository.delete(lastScene);
-        return ApiResponseUtil.success(lastScene, "투표 결과에 따라 삭제됨", HttpStatus.OK, request.getRequestURI());
+
+        if(!deleteSceneRequest.isAccepted()){
+
+            //사용한 카드 삭제해야함
+            PlayerStatus playerStatus = gameRepository.getPlayerStatus(deleteSceneRequest.getGameId(), deleteSceneRequest.getUserId());
+
+            StoryCard storyCard = playerStatus.getStoryCards().stream().filter(storyCard1 -> storyCard1.getId() == deleteSceneRequest.getCardId()).findFirst().orElse(null);
+
+            playerStatus.getStoryCards().remove(storyCard);
+
+            //sceene 데이터 삭제
+            SceneRedis lastScene = scenes.get(scenes.size() - 1);
+            redisSceneRepository.delete(lastScene);
+            return ApiResponseUtil.success(lastScene, "투표 결과에 따라 삭제됨", HttpStatus.OK, request.getRequestURI());
+        }
+
+        return ApiResponseUtil.failure("투표 결과 찬성으로 삭제되지 않음",HttpStatus.CONFLICT,request.getRequestURI());
+
     }
 
     public ResponseEntity<?> getScenesByGameId(String gameId, HttpServletRequest request) {
